Polaris: A KiCad plugin to manage part placement
Sometimes it is nice to specify the position of footprints in a
declarative way.  Polaris allows this by providing a simple command
language to specify a sequence of translations and rotations for a part
in the parts properties.

Eventually Polaris could provide a GUI to support generating these
command sequences, as well as footprints for grouping parts and for
use as origins for the groups.